Refactor and add documentation to many functions.
